redux fix, refactor
