+++ client/src/action/alert.js
@@ -1,13 +1,10 @@
-import {v4 as uuidv4} from "uuid";
-import {SET_ALERT, REMOVE_ALERT} from "./constants";
+import {addAlert, removeAlert} from "../reducers/alert";
 
-export const setAlert =
-  (msg, alertType, timeout = 5000) =>
-  (dispatch) => {
-    const id = uuidv4();
-    dispatch({
-      type: SET_ALERT,
-      payload: {msg, alertType, id},
-    });
-    setTimeout(() => dispatch({type: REMOVE_ALERT, payload: id}), timeout);
-  };
+export const showAlert =
+    (alert, timeout = 5000) =>
+        (dispatch) => {
+            dispatch(addAlert(alert));
+            setTimeout(() => {
+                dispatch(removeAlert(alert));
+            }, timeout);
+        };
